cleanup server and client
