+++ client/src/components/Dashboard/FinhubNewsItem.js
@@ -5,18 +5,14 @@ import "./FinhubNewsItem.css";
 
 const FinhubnewsItem = ({ id, headline, summary, url, time }) => {
   let unix_timestamp = time;
-  // Create a new JavaScript Date object based on the timestamp
-  // multiplied by 1000 so that the argument is in milliseconds, not seconds.
-  var date = new Date(unix_timestamp * 1000);
-  // Hours part from the timestamp
-  var hours = date.getHours();
-  // Minutes part from the timestamp
-  var minutes = "0" + date.getMinutes();
-  // Seconds part from the timestamp
-  var seconds = "0" + date.getSeconds();
+
+  let date = new Date(unix_timestamp * 1000);
+  let hours = date.getHours();
+  let minutes = "0" + date.getMinutes();
+  let seconds = "0" + date.getSeconds();
 
   // Will display time in 10:30:23 format
-  var formattedTime =
+  let formattedTime =
     hours + ":" + minutes.substr(-2) + ":" + seconds.substr(-2);
 
   return (
