+++ client/src/index.js
@@ -1,14 +1,12 @@
-import "bootstrap/dist/css/bootstrap.min.css";
-// import $ from "jquery";
-// import Popper from "popper.js";
-import "bootstrap/dist/js/bootstrap.bundle.min";
-
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App";
-
 import { BrowserRouter as Router } from "react-router-dom";
+
+import "bootstrap/dist/css/bootstrap.min.css";
+import "bootstrap/dist/js/bootstrap.bundle.min";
+
+import App from "./App";
+import "./index.css";
 
 ReactDOM.render(
   <React.StrictMode>
